requirements.txt
google == 3.0.0
pyttsx3 == 2.90
googlesearch-python == 2020.0.2
datetime
PyAudio == 0.2.11
subprocess.run == 0.0.8
SpeechRecognition == 3.6.3
wikipedia == 1.4.0